Grab a random item from an array
Since array indexes start with 0, the last item in the array is always
array.length - 1.

We cannot use Math.round in this case as the largest number it will
return us is array.length.

Math.floor will return us a number less than or equal to the floating
point numbers passed in to it. So the largest number it will return us
is array.length - 1.